Move database objects to DatabaseModels
